Flatten coordinates metadata when converting element dict to csv
